requirements.txt
streamlit
google-generativeai
sentence-transformers
scikit-learn
numpy
requests
networkx
pyvis
neo4j
python-dotenv